style(backend): repository 파일생성

+++ Backend/Archiview/src/main/java/com/ssafy/archiview/repository/RecruitRepository.java
@@ -4,5 +4,5 @@ import com.####.archiview.entity.Recruit;
 import com.####.archiview.entity.User;
 import org.springframework.data.jpa.repository.JpaRepository;
 
-public interface RecruitRepository extends JpaRepository<Recruit, Long> {
+public interface RecruitRepository extends JpaRepository<Recruit, Integer> {
 }
